Terminate before sending timed out /solve request (#3291)
# Description
Sometimes the driver only gets to the point of sending the `/solve`
request to the solver **after** the deadline has passed. This can happen
when the driver's pre-processing takes an unusual amount of time.
Of course we should investigate how we can prevent these outliers from
happening at all but having this safe guard makes sense regardless.

# Changes
Don't send `/solve` request to the solver if the deadline has already
been exceeded.