New contracts for flashloan (#3292)
# Description
Federico published a new PR that supports both Maker and Aave lenders:
https://github.com/cowprotocol/flash-loan-wrapper-solver/pull/4

The idea behind it is that a driver should use `IFlashLoanSolverWrapper`
for all types of borrowers, meaning this contract abstracts all common
stuff that the driver is supposed to know about when encoding bytes
needed for flash loaning.

The one thing that remains a question is a flash fee that is taken by
lender. In case of Maker, we will assume this is 0. In case of Aave, we
need to ask for a fee and that's why we import `IAavePool`. This is the
bare minimum of contracts needed to do the encoding.

Besides that, also added `AaveFlashLoanSolverWrapper` and a new version
of `ERC3156FlashLoanSolverWrapper` which are needed for e2e tests.

Contracts are generated in the same way as in
https://github.com/cowprotocol/services/pull/3268.

<!--
## Related Issues

Fixes #
-->