Store settlement execution events (#3289)
# Description
In order to improve the experience of oncall investigation related to
MEVBlocker submissions and other settlement tx failures, a suggestion is
to implement a CI job that will be able to collect auction execution
stats, such as settlement execution start/end timestamps/blocks,
addresses of the winning solvers, and a result explaining the outcome.
This data is expected to be merged with onchain data such as tx hashes,
block number attempts and builder addresses. That should save a lot of
time on tx expiration/revert event investigations.

# Changes

- [ ] A new DB table where each row represents the minimum required data
for each settlement.
- [ ] Each row is populated in 2 steps: when the settlement execution is
just started and then updates the existing row with the remaining data
when it finishes.
- [ ] A composite PK `(auction_id, solver)` is used since each auction
might have multiple winners(leading to multiple settlement txs).
- [ ] All the errors related to the execution stats storing are just
logged since they should not block the runloop.

## How to test
New DB tests.