Start using `solvers_dto` in the `driver::infra::dto` (#3296)
# Description
I recently noticed some minor discrepancies between `solvers_dto` crate
and `driver::infra::dto` module. Also, we update `solver_dto`
separately, and sometimes, it gets forgotten. The proposal is to start
using `solvers_dto` in the `driver::infra` to avoid any potential
discrepancies in the future, and it should be easier to maintain it.

## How to test
Existing tests should be sufficient.