[EASY] Min won auctions participation guard threshold (#3309)
Currently, the db-statistics-based solver participation guard might ban
a solver that won only 1 or 2 times in the last 100 auctions and failed
to settle them. This is a very sensitive threshold, where the suggestion
is to start considering the solver as low-settling that won at least 3
auctions among 100 latest. The value is configurable and currently
equals the non-settling guard default config.