Refresh gas price before settle executions (#3312)
# Description
Currently `/settle` function of the `driver` is using potentially
outdated gas price to settle solutions onchain. This gas price was
determined at the solving time.

Now we refresh the gas price just before we process settle, so the
transaction is submitted considering the latest state of the gas price.

# Changes
<!-- List of detailed changes (how the change is accomplished) -->

- [ ] Refresh the gas price just before the settle is called.

## How to test
None

<!--
## Related Issues

Fixes #
-->