Single flashloan encoding (#3279)
# Description
Tackles https://github.com/cowprotocol/services/issues/3218

Implements on flavor of the "repay debt with collateral" idea an a e2e
test. The trader is a gnosis safe to make it easier to do all the
necessary pre-interactions. The basic flow is like this:
1. make trader a safe
2. deposit 50K USDC
3. borrow 1 WETH
4. build appdata
    * flashloan 1 WETH (`flashloan` hint)
    * repay 1 WETH (1st pre-hook, tx pre-signed by safe)
    * withdraw 50K USDC (2nd pre-hook, tx pre-signed by safe)
5. place order with that appdata like:
```json
{
    "from": "<SAFE_ADDRESS>",
    "receiver": "<SETTLEMENT_ADDRESS>",
    "sell": "50_000 USDC",
    "buy": "1 WETH + flashloanFee",
    "kind": "buy",
}
```
6. assert that trade got executed, indexed and that changed balances
make sense
    

This flavor of "repay debt with collateral" leaves trace amounts of debt
in the AAVE pool. The alternative would be either over paying slightly
or a bunch of very complex accounting logic using another helper
contract. Both are more involved and not necessary to proof the concept.

# Changes
some bug fixes / workarounds that were necessary to make it work:
* introduce exceptions in filtering logic based on balances in autopilot
and driver to prevent the order from getting filtered out
* forked e2e test setup currently does some things that will not needed
once the flashloan contracts are actually deployed:
    * deploy contracts
    * allow list them as solvers in the settlement contract
* code to pipe the flashloan data around

# Things not addressed by this PR
* proper configuration options in the driver
* more convenience for computing flashloan amounts that need fees
* currently the fee gets added on top of the flashloan amount which
means the user needs to be aware that their `buyAmount` needs to be
`appdata.flashloan.amount + fee`. Ideally the driver would reduce the
flashloan amount such that `loanedAmount + fee = totalSignedAmount`.
* error handling
* supporting multiple flashloans in the same settlement

## How to test
added new forked e2e test
adjusted unit test for solvable orders filtering exception for flashloan
orders

---------

Co-authored-by: MartinquaXD <martin.beckmann@protonmail.com>