Add config for native price estimator in driver (#3316)
# Description
This PR adds configuration for 3 parameters that are used for
calculation of priority fee in the native gas price estimator.
Note that the default values are equal to currently hardcoded values in
`gas_estimation` crate, so no change in behaviour is expected when PR is
merged.

I've tested the native gas price estimator on Reth node, and IMO this
estimator is indeed usable, however, these configuration parameters
currently have bad values - only 20% of the lowest paid transactions are
considered for analysis, which is the main factor why suggested gas
prices are so low, considering the current gas prices used on mainnet.

When this value is increased to 50%, and other two parameters to 50% -
80%, we get much more appropriate prices, with time limits (0s, 18s,
30s):

```
{"base_fee_per_gas":0.707347445,"max_fee_per_gas":2.414698008,"max_priority_fee_per_gas":0.500001559}
{"base_fee_per_gas":0.707347445,"max_fee_per_gas":1.6323970757835722,"max_priority_fee_per_gas":0.15583408436011478}
{"base_fee_per_gas":0.707347445,"max_fee_per_gas":1.5986310602835723,"max_priority_fee_per_gas":0.1386735564680693}
```

These new configuration values practically mean that we don't want to
suggest prices that barely bring our transaction onchain, but instead,
we want our transaction to aim to be averagely paid (we can even aim to
be among highest paid transactions with these parameters).

I will test this and other configuration sets to make sure the values
are also meaningful in volatile periods.

IMO, when these parameters are properly set, we will probably be able to
completely drop additionalTip and minimalTip, as those params were
introduced to adjust the native gas price estimator in a simplest way
possible, without having to dig into the complex implementation of the
estimator.

# Changes
<!-- List of detailed changes (how the change is accomplished) -->

- [ ] Added configuration for NativeGasPriceEstimator

## How to test
<!--- Include details of how to test your changes, including any
pre-requisites. If no unit tests are included, please explain Testing
locally Native gas price estimator with different configuration sets.

<!--
## Related Issues

Fixes #
-->