driver: get access list for balance calculation (#3300)
# Description
From the task:
This
[order](https://explorer.cow.fi/gc/orders/0x9955075846884c8a992b5d61bdb8cc11b4049fe23ebda2420d7f8e0599364ecf5a945741f5c367410678e1513b939f7adafbacf067917f5d?tab=overview)
was filled sub-optimally in 2 partial fills.
It contains a pre-hook that claims rewards to provide the full sell
token balance to the owner. The driver already computes the available
balances for an order based on the outcome of the pre-hooks (if there
are any). But in this case the balance simulation didn't show that the
owner would get all the necessary sell tokens from the hook so it
reported the available balance to be whatever the owner currently has
without the hook.
A solver picked up this order and executed it. In the actual settlement
the hook got executed and unlocked the remaining sell tokens which led
to another partial fill that then fulfilled the order completely.

When simulating the hook on its own it exceeds its gas limit and
therefor reverts. This explains why no balance change happens in the
simulation. However, in the actual settlement the balance change
happened.
When comparing the [hook
simulation](https://dashboard.tenderly.co/cow-protocol/solver-slippage/simulator/460b7ac7-6935-4800-9034-80cae63d5490)
with the onchain settlement
[simulation](https://www.tdly.co/tx/100/0xadf0c46fee5d5e728eb8959bb411b92ff36e44b6eb150ec911ddee9dd95ef538)
it became apparent that executing the hook as part of the actual
settlement was a lot cheaper.

The reason seems to be that the storage loads are a lot cheaper in the
real settlement. That is likely caused by the fact that we use access
lists for the real solution but not during the balance simulation.

# Changes
- Get the access list before simulating the balance

## How to test
1. Regression tests

## Related Issues
Fixes #3252