Return error on order replacement when it's not safe (#3323)
# Description
We want to allow users to "edit" orders, implemented by canceling a
previous order when creating a new one. We have a `replacedOrder`
property for this.

One concern is that this could lead to double spending when the old
order is part of the current auction and the winning solver intends to
settle it.

The goal of this PR is to mitigate this issue as much as possible by
checking whether any solver has bid on the original order in the last
few auctions. The POST `/orders` method now returns a `400 Bad Request`
in that case.

A follow-up PR can extend this logic to the cancellation endpoint. In
this case, instead of rejecting the cancellation outright, we may prefer
to return a warning message instead.

# Changes
- Added a new database operation `load_latest_competitions` to fetch all
the competitions from the last N auctions.
- Updated order replacement logic in the orderbook to verify that the
order is not included in any recent competitions (using the new
`order_is_actively_bid_on` function).
- Introduced a more granular order replacement error specification via
the new `OrderReplacementError` enum.
- Updated `openapi.yml` specification with a new error variant in
`OrderPostError`.
- New orderbook argument `active_order_competition_threshold`.

## How to test
Created a new e2e scenario, `try_replace_active_order_test`, under the
`replace_order` e2e module, which triggers the new error.

However, the ideal scenario would be to test this when the old order is
included in any solution by using a mock solver that always returns a
bad solution containing the old order. However, achieving this would
require the ability to mock the driver in e2e tests, which we currently
lack. A follow-up PR can address this.

## Related Issues

Fixes https://github.com/cowprotocol/services/issues/3315