Save blocks passed since first submission attempt (#3322)
# Description
New metric that saves the number of blocks passed between the first
submission attempt and the block at which the error is detected.

This metric should tell us more whether the error is caused by
MEVBlocker being down and not processing transactions (in this case we
would expect high value for blocks passed) or it's caused by high price
volatility which should result in revert in one or two blocks just after
the submission started (in most cases).

---------

Co-authored-by: MartinquaXD <martin.beckmann@protonmail.com>