Add support for multiple partner fees (#3320)
# Description
Add support for multiple partner fees.

# Changes
- Support for multiple partner fees
- Backwards compatibility
- The partner fees are squashed into one partner fee, so it is easier to
apply the max cap

## How to test
1. Unit tests

NOTE: Because of the partner fees being squashed into one and
abstracting that to a single function for better unit testing, I saw no
benefit of adding an e2e to it, but I am open to it if you see it fit.

---------

Co-authored-by: MartinquaXD <martin.beckmann@protonmail.com>