Fix flashloan default lender (#3330)
# Description
This PR makes the driver configuration for flashloans more
production-ready.

Using MAKER(SKY) address as hardcoded default lender doesn't make sense
anymore, because SKY only lends DAI token. If, instead, we wanted to
replace it with AAVE or some other lender that lends all sorts of
tokens, then it means this lender has to be deployed on all networks we
support.

With that, it makes most sense to not have a hardcoded default lender
and just define it in the infra, which should avoid weird bugs.

# Changes
<!-- List of detailed changes (how the change is accomplished) -->

- [ ] Removed default lender config in the driver

## How to test
Existing tests.

<!--
## Related Issues

Fixes #
-->