[EASY] Log order quote's solver (#3329)
# Description
As a first step to improve the quote competition visualization for a
specific order, this PR updates the order creation log by including the
selected quote solver address, which eliminates the redundant step of
searching for this data in either logs or db. If the initiative gets
approved, this can be improved by logging the solver's name. However,
this would require additional work to build a mapping between the
solver's address and its name.

## How to test
Observe the logs.