orderbook: Return 404 for get order status when competition was not found (#3328)
# Description
Right now we return internal error when the solver competition was not
found.

Return `NOT_FOUND` (404) error instead of internal error (500) when
there solver competition was not found for the given order.

## How to test
1. Regression test

The change is fairly simple, I assume adding an e2e test for this is a
little bit of an overkill, but I am open to the idea too.

---------

Co-authored-by: MartinquaXD <martin.beckmann@protonmail.com>