Hide quote execution plan (#3354)
# Description
Currently we always expose the metadata of an order's quote. This
basically contains the execution plan used to generate the quote. If
this is made public while an order is still fillable you could
theoretically have a solver that just tries to copy the quote's
execution plan. That could lead to quoters not providing their best
solution because they are afraid that the copy solver would snipe their
order.
If we only show the execution plan after an order is no longer fillable
this strategy will no longer work and the incentives of the quote
competition remain unchanged (compared to never showing the execution
plan like we did before).

# Changes
We now hide a quote's `metadata` until an order is no longer fillable.

## How to test
Adjusted an existing e2e test