[PoC] Improve price estimation error (#3363)
# Description
Currently, it's quite difficult to understand why a quoting competition
failed. All the estimators use the same error struct and much more time
than anticipated needs to be spent to find the cause of the problem.

The issue that inspired me to open this PR was the unstable quoting
results for a particular token pair. If I filter the logs by this pair,
with this PR, I will easily see the list of all the errors and can
decide what the majority of the causes are. With any extra step, this
won't be possible(as right now). I would need to get the request ID each
time. This is a very annoying problem in other backend components, which
is much harder/impossible to address.

# Changes

- [ ] As a first step, the price estimation error is enriched with the
estimator kind that caused the competition to fail.
- [ ] Using this, it should be much easier to understand at which step
the competition has failed (e.g., whether a Native sell token estimator
or a regular one failed, and then dive deeper into the exact problem if
needed).
- [ ] It doesn't change the API, only the logging.

## How to test
Observe logs.

Sampled log:
```
2025-04-25T10:23:39.841Z  WARN request{id="1"}: orderbook::api::post_quote: post_quote error err=error calculating quote: NativeBuy estimator failed: No liquidity request=OrderQuoteRequest { from: 0x2b5ad5c4795c026514f8317c7a215e218dccd6cf, sell_token: 0x5fbdb2315678afecb367f032d93f642f64180aa3, buy_token: 0x68b1d87f95878fe05b998f19b66f4baba5de1aed, receiver: None, side: Sell { sell_amount: AfterFee { value: U256(1000000000000000000) } }, validity: For(1800), app_data: Hash { hash: 0x0000000000000000000000000000000000000000000000000000000000000000 }, sell_token_balance: Erc20, buy_token_balance: Erc20, signing_scheme: Eip712, price_quality: Optimal }
```
- `error calculating quote:`  is an API wrapper
- `NativeBuy estimator failed:`  is a new estimator wrapper
- `No liquidity` - the original error