Reinit subgraph liquidity sources (#3369)
# Description
Currently our subgraph based liquidity sources (balancer v2, uniswap v3)
fetch the list of known pools only on startup. This currently leads to
frequent requests why some pool does not get indexed (e.g. baseline
router is not returning quotes although there is a pool it could use).

# Changes
The optimal solution would be to build our indexing logic such that it
only queries pools that were created since we last checked. However,
that would be pretty involved and overall it doesn't make much sense to
spend much effort on liquidity indexing since none of the competitive
solvers use our indexed liquidity anyway.
To keep the effort to a minimum I adjusted the
`BackgroundInitLiquiditySource` logic such that it can be configured to
periodically rerun the initialization logic. This is effectively
equivalent to restarting the driver manually (in terms of discovering
new pools).

Since using the subgraph costs credits the default `reinit_interval` is
a conservative `12h` but can be overwritten in the driver config file.

## How to test
Added a unit test to verify that the init logic runs multiple times if a
`reinit_interval` was provided.