Surplus partner fee (#3368)
# Description
Implements the ability for partners to specify the `Surplus` and
`PriceImprovement` fee policy (in addition to the already available
`Volume` policy).
The protocol enforces a maximum ratio of the order's total volume that
can be captured via partner fees. Because `Surplus` and `PriceImpact`
fee policies also configure a `max_volume_factor` we are able to cap
these fees as well using the global volume limit.

# Changes
- slight refactor of the autopilot logic that computes the final partner
fee policies (the `PriveImprovement` fee requires a reference quote)
- added logic to apply the new partner fee policies in the autopilot
- added backwards compatible (de)serialization logic for the new partner
fee variants in the `appdata` crate

## How to test
- extended unit tests for new partner fee policy (de)serialization logic
- added new e2e test that verifies that the autopilot attached the
correct fee policies to the order
- also tests the logic to limit the new partner fees to the global
volume cap