Update Fee Amount docs (#3370)
The current description for the fee amount is misleading and outdated.
It's the raw sell token amount. The field is referenced both in the
quote response as well as in the order placement payload.

In the latter it has to be zero as otherwise the order will be rejected
(wo no longer allow for upfront fixed fees).

cc @anxolin