Introduce combinatorial auctions (#3385)
# Description
Implements all details of combinatorial auctions in the
`winner_selection` module. The implementation is already conscious about
performance considerations when computing references scores but to
improve things further we should consider changing the internal data
structures of `domain::Auction` to be more efficient for this from the
get go to avoid converting this data first.
For error handling we drop all solutions where the scores could not be
computed correctly. These successfully computed scores get cached so
that they don't have to be recomputed multiple times.

# Changes

This PR introduces 2 hacky things:
1. It makes some of the `math` module public in order to reuse the
`score` computation. Reusing the score logic makes sense overall but at
the moment I simply made public whatever I needed. This should be
improved with a refactor.
In order to avoid expensive `.clone()` operations I adjusted the
interfaces of the score related functions slightly. Now they don't take
arguments that they don't actually need. Also we can now pass
`fee_policies` as `HashMap<OrderUid, Vec<fee::Policy>` and
`HashMap<OrderUid, &[fee::Policy]>`.
2. I added a new field `computed_score` to the `Solution` which can be
set and read by the combinatorial selection mechanism. The reason is
that the reference score computation should be based on the scores
computed by the autopilot and not by the solver reported value and it
was not really possible to pass it around nicely otherwise. IMO the
proper follow up refactor is to compute the score in the `Solution`
constructor.

## How to test
Reeneabled multi winner test commented out in
https://github.com/cowprotocol/services/pull/3384