`settlements` table clean-up (#3361)
# Description
It turned out that `tx_from` and `tx_nonce` columns in the `settlements`
table have not been populated since the migration was implemented in
https://github.com/cowprotocol/services/pull/2283. Since no one
complained about the absence of this data for more than 1 year, it
probably makes sense to drop these columns since the data still can be
retrieved from the tx itself. There is no reference to these columns in
the code.

# Changes

- [ ] Drop redundant columns

## How to test
Existing tests.

## Related issues

Partially addresses task 4 from #3350