Run participation guard in background (#3386)
# Description
Recently we experienced some serious degradation when we needed to use a
fallback node provider which had high latency. This resulted in up to 4s
between the autopilot cutting an auction and the drivers receiving the
request. The reason was that the participation guard (to check whether a
solver is eligible to participate in the current auction) awaits an RPC
request before we send out the `/solve` request to the driver.
Additionally it looked like these individual RPC requests were not
batched together by our auto batching logic. The end result basically
looked like what you would expect from this code:
```rust
for driver in &drivers {
    if !check_allowed().await {
        anyhow::bail!("solver deny listed");
    }
    tokio::task::spawn(driver.send_request())
}
```

# Changes
There are a few things which we need to handle:
1. send request
2. check if solver is eligible
3. abort request if we exceed the deadline
4. abort request if the solver is not allow listed
5. do everything concurrently
6. handle all the error cases

I tried to make this as readable as possible but I'm not super happy
with the result. LMK if you have ideas how to make this code nicer.