Trivial: log liquidity sources consistently (#3390)
# Description
When logging the number of pools collected for each liquidity source we
collect them into a map to print a small overview.
Currently the order in which these sources get printed is not stable
because we use a `HashMap`.

# Changes
Replace `HashMap` with `BTreeMap` to have a consistent iteration and
therefore logging.