Initialize the settlement contract with deployed function (#3389)
# Description
By using
[`contracts::GPv2Settlement::at`](https://github.com/cowprotocol/services/blob/main/crates/autopilot/src/infra/blockchain/contracts.rs#L30)
the deployment information is not properly detected in [this
line](https://github.com/cowprotocol/services/blob/main/crates/autopilot/src/run.rs#L394),
and it is returning `None`. Changing reading the deployment information
directly from the raw contract since it is built in `build.rs` file.

The reasons why the deployment information is missing by using the call
`at()` is unknown to me. Nevertheless, this information is only needed
for new networks which don't have a last index settlement stored in the
database.

## How to test
I tested it by running it in staging.

---------

Co-authored-by: Martin Magnus <martin.beckmann@protonmail.com>