Add metrics for solve times in driver (#3398)
# Description
Currently we don't have good data on how long solvers have to compute
their solution. We do have a dashboard that shows the remaining time as
observed by the solver engine. While this is the most accurate way to
measure this we also only have data for very few solvers.

# Changes
- measure remaining solve time right before driver starts sending the
request.
- measure how long the solver needed to report back a solution

# Test
- deployed on shadow mainnet and checked the presence of the data with
grafana