[EASY] Fix baseline buy amount calculation (#3402)
# Description
When testing on new networks where the settlement contract buffers are
empty, baseline solver solutions couldn't be verified due to on-chain
reverts that showed a 1 unit higher out amount calculated by the solver
compared to the values returned by liquidity pools. After testing it
with different liquidity pools separately(uniswap and balancer), it
became apparent that there is an issue on the baseline's side.

# Changes
When calculating the buy amount for sell orders, `ceil_div` was used,
probably because the settlement contract uses it. But from the solver's
perspective, it should be a `floor` operation to avoid providing too
good solutions. This PR addresses it.

## How to test
It is hard to validate on the existing networks, but pretty
straightforward on new ones, such as Polygon. No more reverts with the
baseline's solutions. Also, an updated unit test.