Fill flashloan data in driver if solver didn't specify anything (#3416)
# Description
Currently solvers would have to copy the flashloan hints provided in the
auction into their solution to make the driver encode the flashloan
solution correctly. To allow solvers to support flashloans without
having to alter anything in their response the driver actually needs to
copy over the data.

# Changes
With this PR the solvers have 3 options:
* return `flashloans: null` (or don't set the field at all) => driver
will fill in the blanks
* return `flashloans: []` => indicates that the solver fronts the cache
* return `flashloans: [flashloan]` => tells solver exactly which
flashloan to take out (could be the original hints just copied over, or
something completely different)

## How to test
I removed the logic in the baseline solver to actively copy the
flashloan hints into the solution so now it also relies on the driver
doing the copying. The existing flashloan e2e should still pass.