Fetch liquidity at block driver config (#3418)
# Description
An issue was discovered recently on Avalanche where calling
`getReserves()` for non-finalized blocks on UniV2 LP leads to failures
with the `cannot query unfinalized data` error. Avalanche achieves
sub-second finalization, which means it should be safe to use finalized
blocks for fetching liquidity.

# Changes

- [ ] A new driver liquidity configuration allowing to specify which
block will be used for liquidity fetching across all the solvers:
Finalized, Recent, or Latest. The difference between Recent and Latest
is that when using the former, CurrentBlockWatcher is used to get the
latest block, and for the latter, the web3 client decides which block is
the latest.
- [ ] A single config across all the solvers.

## How to test
✅ Tested on Avalanche.

## Alternative solutions

I was also thinking about integrating this config directly into the
liquidity fetcher, in a way that overrides the provided block number in
the `get_liquidity()` function, but this approach seems too implicit and
confusing.