Use new auction abstraction in the shadow autopilot (#3408)
# Description
In order to quickly get confidence that the combinatorial auction logic
is correct we'd like to run it in the shadow environment as well (also
it's just nicer to make shadow and regular behave as same as possible).
This PR rips out some of the shadow specific code and replaces it with
code used in the regular autopilot wherever possible.

# Changes
- use new `Arbitrator` abstraction in shadow autopilot
- allow solvers to return multiple solutions
- simplify some of the error handling (according to the solver team they
only check how often solvers throw errors but not which in particular)
- kept logs about winners, rewards, and calldata around (solver team
said this is important)

## How to test
Already deployed it for a while in the cluster and confirmed with the
solver team that all the logs / data they need are still there.