Simplify ethflow indexing start block search (#3414)
# Description
Currently, determining the start block of indexing ethflow events on
fresh chains with no corresponding orders/refunds requires using an RPC
node that has access to either the ethflow or settlement contract
deployment blocks.

When dealing with pruned full nodes as a primary transport, a solution
for that could be to use an archive node for the initial start that will
index all the old blocks, and then switch back to the pruned one. With
the current implementation, that is not possible, which still requires
access to either the ethflow db data or the contract deployment block.
Another ethflow indexing start block or even "skip indexing" flag can be
provided via the config to solve this, but the autopilot's logic remains
outdated.

With #3075, this logic can be simplified by relying on the last indexed
blocks stored in the corresponding DB table.

# Changes

1. Search for a corresponding record in the `last_indexed_blocks`. Use
the next block as the starting index.
2. In case it doesn't exist yet or the index is 0, use the configured
ethflow indexing start block value.
3. Fallback to the settlement contract deployment block number.
4. Use the determined number to fetch the block using the connected RPC
node to ensure it is able to continue indexing.

With this logic, the pruned node can safely be used on networks with no
ethflow orders yet, but with already indexed data.

## How to test
✅ Tested on Avalanche + pruned full node, which fails to start with the
`main` services image but works with this PR.

## Related PRs

Simplifies #2931, utilized #3075