[TRIVIAL] Improve tables size query performance (#3422)
# Description
I recently noticed that updating metrics for some of the tables has
started consuming a lot of time(1-5s). This PR moves the problematic
tables from the `COUNT (*) FROM` query to the optimized version, which
is being used with some large tables. As a result, there is a chance of
not receiving an accurate row count, but it shouldn't be a problem.

## How to test
Observe logs.

## Follow-ups

The following query is also not efficient anymore. This should be
addressed in a follow-up PR.

https://github.com/cowprotocol/services/blob/ca7b8b04ff5b1138cb99418d0116a802b9d9cf0f/crates/autopilot/src/database/mod.rs#L92-L104