Combinatorial auctions cutover (#3423)
# Description
We recently started implementing combinatorial auctions in the autopilot
[[1]](https://github.com/cowprotocol/services/pull/3384)[[2]](https://github.com/cowprotocol/services/pull/3385).

Right now the use of the new combinatorial auction logic is controlled
by the `max_winners_per_auction` configuration parameter. If it's
greater than `1` the autopilot will use combinatorial auctions.

But the solver team requested to be able to enable the use of
combinatorial auctions based on a cutover timestamp. So when the
configured cutover timestamp arrives, the autopilot should automatically
switch to the combinatorial auctions.

This PR adds a cutover argument in the autopilot. 

Once combinatorial auctions are in place we can remove the cutover and
clean up the related logic
(https://github.com/cowprotocol/services/issues/3424).

# Changes
<!-- List of detailed changes (how the change is accomplished) -->

- [ ] New optional autopilot argument`combinatorial_auctions_cutover`
- [ ] Updated the logic of the `single_winner` function (which decides
the arbitrator to use):
- [ ] Single winner if the auction deadline is not past the
`combinatorial_auctions_cutover` or `max_winners_per_auction` is `1`,
combinatorial auctions otherwise
- [ ] This setup allows us to have a `max_winners_per_auction` value > 1
set up beforehand without triggering combinatorial auctions until the
cutover.
- [ ] The arbitrator is built again for each auction, as the
implementation to use is now decided based on the auction deadline.
- [ ] Updated the e2e setup with a past cutover so the multiple winner
test use combinatorial auctions

## How to test
This can only be properly tested on staging environment by setting a
cutover. For now the existing tests pass.

## Related Issues
- [ ] https://github.com/cowprotocol/services/issues/3387
- [ ] https://github.com/cowprotocol/services/issues/3424